V9 oJO que cuit y dni en persona ahora son Int64

+++ TPC_Ayala_Herrera/Dominio/Persona.cs
@@ -11,8 +11,8 @@ namespace Dominio
         public int Id { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int Dni { get; set; }
-        public int Cuit { get; set; }
+        public Int64 Dni { get; set; }
+        public Int64 Cuit { get; set; }
         public string Domicilio { get; set; }
         public string Mail { get; set; }
         public string Telefono { get; set; }
